Refactor `lookupNetworkTests` helper function
The `lookupNetworkTests` helper function has been updated to
distinguish between the initial state and the expected provider config
at the end of the operation under test.

Previously the `providerConfig` passed in was used as part of the
initial state _and_ as the expected end state. This worked for
operations like `resetConnection` and `lookupNetwork` where the network
doesn't change, but it was incompatible with functions that change the
active network.

Relates to #1210